Rename batch_sys_handlers -> job_runner_handlers
