+++ tests/functional/job-file-trap/00-sigusr1/python/background_vacation.py
@@ -16,7 +16,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-from cylc.flow.batch_sys_handlers.background import BgCommandHandler
+from cylc.flow.job_runner_handlers.background import BgCommandHandler
 
 
 class MyBgCommandHandler(BgCommandHandler):
