+++ tests/functional/job-submission/00-user/python/my_background.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from cylc.flow.batch_sys_handlers.background import BgCommandHandler
+from cylc.flow.job_runner_handlers.background import BgCommandHandler
 
 
 class MyBgCommandHandler(BgCommandHandler):
