+++ tests/unit/job_runner_handlers/test_slurm.py
@@ -17,7 +17,7 @@
 import pytest
 import os
 
-from cylc.flow.batch_sys_handlers.slurm import JOB_RUNNER_HANDLER
+from cylc.flow.job_runner_handlers.slurm import JOB_RUNNER_HANDLER
 
 home = os.path.expandvars('$HOME/')
